Simple algorithm

+++ api/src/main/java/com/bigdata/hadup/graphmovierecommendationengine/model/Movie.java
@@ -1,10 +1,8 @@
 package com.bigdata.hadup.graphmovierecommendationengine.model;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import org.neo4j.ogm.annotation.*;
 
 import java.io.Serializable;
-import java.util.HashSet;
 import java.util.Objects;
 import java.util.Set;
 
@@ -25,7 +23,7 @@ public class Movie implements Serializable {
     private Integer length;
 
     @Property("numVotes")
-    private Integer numVotes;
+    private Long numVotes;
 
     @Property("rating")
     private Double rating;
@@ -90,11 +88,11 @@ public class Movie implements Serializable {
         this.length = length;
     }
 
-    public Integer getNumVotes() {
+    public Long getNumVotes() {
         return numVotes;
     }
 
-    public void setNumVotes(Integer numVotes) {
+    public void setNumVotes(Long numVotes) {
         this.numVotes = numVotes;
     }
